requirements.txt
pandas>=2.2.0
numpy>=1.26.0,<2.0
scipy>=1.13.0
scikit-learn>=1.5.0
statsmodels>=0.14.0
matplotlib>=3.9.0
seaborn>=0.13.0
polars>=1.17.0
pyarrow>=16.0.0
pyyaml>=6.0
geopandas>=1.0.0
networkx>=3.3
lightgbm>=4.5.0
shap>=0.46.0
ruptures>=1.1.9
pyod>=2.0.0
powerlaw>=1.5.0
tqdm>=4.67.0
pytest>=7.4.0
joblib>=1.3.0

# Phase 6: Advanced Analysis
torch>=2.2.0
xgboost>=2.0.0
catboost>=1.2.0
umap-learn>=0.5.5
hdbscan>=0.8.33
tslearn>=0.6.3
optuna>=3.6.0
python-louvain>=0.16

# Phase 7: SOTA Analysis
ripser>=0.6.0
persim>=0.3.0
pot>=0.9.0
econml>=0.15.0
